Share delta-rule validation and packed references
## Human Note

## Agent note

Extract the structural Q/K/V, gate, beta, packed-offset, V-major state-shape, and device contract
shared by KDA and GDN while leaving each family responsible for its own gate layout and dtype
policy. Reuse one eager packed-reference loop for both families, preserving their different compute
dtypes and dense-operation arguments. Chunk metadata and scheduler ownership remain unchanged.

## Test Plan

```bash
pytest -n 6 -q test/test_kda_recurrent.py test/test_kda.py test/linear/test_gdn.py test/linear/test_gdn_fused.py test/test_kda_impl_dispatch.py
ruff check attn_gym/linear/_delta_rule/validation.py attn_gym/linear/_delta_rule/reference.py attn_gym/linear/kda/validation.py attn_gym/linear/gdn/validation.py attn_gym/linear/kda/impl/reference.py attn_gym/linear/gdn/impl/reference.py attn_gym/linear/gdn/api.py
```

stack-info: PR: https://github.com/meta-pytorch/attention-gym/pull/400, branch: drisspg/stack/108
