+++ attn_gym/linear/gdn/impl/reference.py
@@ -2,68 +2,12 @@
 
 from __future__ import annotations
 
-from itertools import pairwise
-
 import torch
 import torch.nn.functional as F
 
+from attn_gym.linear._delta_rule.reference import packed_delta_rule_reference
+
 _CHUNK_SIZE = 64
-
-
-def _packed_reference(
-    dense_op,
-    q: torch.Tensor,
-    k: torch.Tensor,
-    v: torch.Tensor,
-    log_decay: torch.Tensor,
-    beta: torch.Tensor,
-    initial_state: torch.Tensor | None,
-    cu_seqlens: torch.Tensor,
-    scale: float,
-    output_final_state: bool,
-) -> tuple[torch.Tensor, torch.Tensor | None]:
-    """Evaluate packed sequences independently through a dense reference operation."""
-    heads, key_dim, value_dim = q.shape[2], q.shape[3], v.shape[-1]
-    num_sequences = cu_seqlens.shape[0] - 1
-    output = torch.zeros_like(v)
-    final_state = None
-    if output_final_state:
-        final_state = (
-            q.new_zeros(num_sequences, heads, value_dim, key_dim)
-            if initial_state is None
-            else initial_state.clone()
-        )
-
-    offsets = cu_seqlens.cpu().tolist()
-    if (
-        offsets[0] != 0
-        or any(begin > end for begin, end in pairwise(offsets))
-        or offsets[-1] > q.shape[1]
-    ):
-        raise ValueError(
-            "cu_seqlens offsets must start at zero, be nondecreasing, and end within "
-            "the physical token capacity"
-        )
-    for sequence, (begin, end) in enumerate(pairwise(offsets)):
-        if begin == end:
-            continue
-        span = slice(begin, end)
-        span_output, span_state = dense_op(
-            q[:, span],
-            k[:, span],
-            v[:, span],
-            log_decay[:, span],
-            beta[:, span],
-            scale=scale,
-            initial_state=None
-            if initial_state is None
-            else initial_state[sequence : sequence + 1],
-            output_final_state=output_final_state,
-        )
-        output[:, span] = span_output
-        if final_state is not None:
-            final_state[sequence] = span_state[0]
-    return output, final_state
 
 
 def reference_gdn(
@@ -101,7 +45,7 @@ def reference_gdn(
                 output_final_state=output_final_state,
             )
         else:
-            output, state = _packed_reference(
+            output, state = packed_delta_rule_reference(
                 dense_op,
                 q,
                 k,
@@ -110,8 +54,8 @@ def reference_gdn(
                 beta,
                 initial_state,
                 cu_seqlens,
-                scale,
                 output_final_state,
+                scale=scale,
             )
     return output.to(output_dtype), state
